Pass SurfaceFrame to SubmitFrame
