+++ shell/common/shell_test_external_view_embedder.cc
@@ -33,13 +33,11 @@ SkCanvas* ShellTestExternalViewEmbedder::CompositeEmbeddedView(int view_id) {
 }
 
 // |ExternalViewEmbedder|
-bool ShellTestExternalViewEmbedder::SubmitFrame(GrContext* context,
-                                                SkCanvas* background_canvas) {
-  return true;
+bool ShellTestExternalViewEmbedder::SubmitFrame(
+    GrContext* context,
+    std::unique_ptr<SurfaceFrame> frame) {
+  return frame->Submit();
 }
-
-// |ExternalViewEmbedder|
-void ShellTestExternalViewEmbedder::FinishFrame() {}
 
 // |ExternalViewEmbedder|
 void ShellTestExternalViewEmbedder::EndFrame(
