+++ shell/platform/android/external_view_embedder/external_view_embedder.cc
@@ -40,17 +40,18 @@ std::vector<SkCanvas*> AndroidExternalViewEmbedder::GetCurrentCanvases() {
 }
 
 // |ExternalViewEmbedder|
-bool AndroidExternalViewEmbedder::SubmitFrame(GrContext* context,
-                                              SkCanvas* background_canvas) {
+bool AndroidExternalViewEmbedder::SubmitFrame(
+    GrContext* context,
+    std::unique_ptr<SurfaceFrame> frame) {
   // TODO(egarciad): Implement hybrid composition.
   // https://github.com/flutter/flutter/issues/55270
   TRACE_EVENT0("flutter", "AndroidExternalViewEmbedder::SubmitFrame");
   for (size_t i = 0; i < composition_order_.size(); i++) {
     int64_t view_id = composition_order_[i];
-    background_canvas->drawPicture(
+    frame->SkiaCanvas()->drawPicture(
         picture_recorders_[view_id]->finishRecordingAsPicture());
   }
-  return true;
+  return frame->Submit();
 }
 
 // |ExternalViewEmbedder|
@@ -80,11 +81,6 @@ void AndroidExternalViewEmbedder::ClearFrame() {
 
 // |ExternalViewEmbedder|
 void AndroidExternalViewEmbedder::CancelFrame() {
-  ClearFrame();
-}
-
-// |ExternalViewEmbedder|
-void AndroidExternalViewEmbedder::FinishFrame() {
   ClearFrame();
 }
