+++ shell/platform/embedder/embedder_external_view_embedder.cc
@@ -129,8 +129,9 @@ static FlutterBackingStoreConfig MakeBackingStoreConfig(
 }
 
 // |ExternalViewEmbedder|
-bool EmbedderExternalViewEmbedder::SubmitFrame(GrContext* context,
-                                               SkCanvas* background_canvas) {
+bool EmbedderExternalViewEmbedder::SubmitFrame(
+    GrContext* context,
+    std::unique_ptr<SurfaceFrame> frame) {
   auto [matched_render_targets, pending_keys] =
       render_target_cache_.GetExistingTargetsInCache(pending_views_);
 
@@ -263,10 +264,7 @@ bool EmbedderExternalViewEmbedder::SubmitFrame(GrContext* context,
                                            std::move(render_target.second));
   }
 
-  return true;
+  return frame->Submit();
 }
-
-// |ExternalViewEmbedder|
-void EmbedderExternalViewEmbedder::FinishFrame() {}
 
 }  // namespace flutter
